division of the front on user and admin pabels
