+++ src/webapp/src/api/user.js
@@ -1,4 +1,4 @@
-import config from '../../../config/dev'
+// import config from '../../../config/dev'
 
 export default {
   //
