+++ src/webapp/src/components/Admin/Users.vue
@@ -1,0 +1,15 @@
+<template>
+    <div>
+        Users
+    </div>
+</template>
+
+<script>
+export default {
+    name: 'AdminUsers'
+}
+</script>
+
+<style scoped>
+
+</style>
